@@ -9,7 +9,7 @@ RSpec.describe ProjectsController, type: :controller do
     end
 
     it "responds successfully with an HTTP 200 status code" do
-      expect(response).to be_success
+      expect(response).to be_successful
       expect(response).to have_http_status(200)
     end
 
@@ -31,7 +31,7 @@ RSpec.describe ProjectsController, type: :controller do
       end
 
       it "responds successfully with an HTTP 200 status code" do
-        expect(response).to be_success
+        expect(response).to be_successful
         expect(response).to have_http_status(200)
       end
 
@@ -50,7 +50,7 @@ RSpec.describe ProjectsController, type: :controller do
       end
 
       it "responds successfully with an HTTP 200 status code" do
-        expect(response).to be_success
+        expect(response).to be_successful
         expect(response).to have_http_status(200)
       end
 
