@@ -4,7 +4,7 @@ RSpec.describe SearchController, type: :controller do
   describe 'GET #index' do
     it 'responds successfully with an HTTP 200 status code' do
       get :index
-      expect(response).to be_success
+      expect(response).to be_successful
       expect(response).to have_http_status(200)
     end
 
